feat(many): add a way for consumers to pass their own component theme definition so they can build their own components with InstUI
