+++ packages/emotion/src/useStyle.ts
@@ -23,101 +23,70 @@
  */
 
 import { useTheme } from './useTheme'
-import { mergeDeep } from '@instructure/ui-utils'
-import type {
-  NewComponentTypes,
-  SharedTokens,
-  Theme
-} from '@instructure/ui-themes'
+import { getComponentThemeOverride } from './getComponentThemeOverride'
+import type { ComponentTheme } from '@instructure/shared-types'
+import type { Theme } from '@instructure/ui-themes'
 
 // returns the second parameter of a function
 type SecondParameter<T extends (...args: any) => any> =
   Parameters<T>[1] extends undefined ? never : Parameters<T>[1]
 
-type NewThemeOverrideProp<
-  ComponentTheme extends ReturnType<NewComponentTypes[keyof NewComponentTypes]>
+type GenerateComponentTheme = (theme: Theme) => ComponentTheme
+
+type UseStyleParamsWithTheme<
+  P extends (componentTheme: any, params: any, theme: any) => any
 > = {
-  themeOverride?:
-    | Partial<ComponentTheme>
-    | ((compTheme: ComponentTheme, theme: Theme) => Partial<ComponentTheme>)
+  generateStyle: P
+  params?: SecondParameter<P>
+  generateComponentTheme: GenerateComponentTheme
+  componentId: string
+  displayName?: string
 }
 
-type GenerateStyleFn<ComponentTheme> =
-  | ((
-      componentTheme: ComponentTheme,
-      params: any,
-      sharedTokens: SharedTokens
-    ) => any)
-  | ((componentTheme: ComponentTheme, params: any) => any)
-  | ((componentTheme: ComponentTheme) => any)
+// TODO this is only used by the old themes, remove when everything uses the new
+// theming system
+type UseStyleParamsWithoutTheme<
+  P extends (componentTheme: any, params: any, theme: any) => any
+> = {
+  generateStyle: P
+  params?: SecondParameter<P>
+  generateComponentTheme?: undefined
+  componentId?: undefined
+  displayName?: undefined
+}
 
-/**
- * new useStyle syntax, use this with v11.7+ themes
+/*
+ * This is only used by the **old themes**, remove when everything uses the new
+ * theming system (InstUI v12)
  */
-// TODO: improve useStyle to handle generateStyle functions that don't
-// have a theme.
 const useStyle = <
-  ComponentTheme extends ReturnType<NewComponentTypes[keyof NewComponentTypes]>,
-  GenerateStyle extends GenerateStyleFn<ComponentTheme>
->(useStyleParams: {
-  generateStyle: GenerateStyle
-  params?: SecondParameter<GenerateStyle>
-  // needs to be a string too because it might be a child component
-  componentId: keyof NewComponentTypes | string
-  themeOverride?: NewThemeOverrideProp<ComponentTheme>['themeOverride']
-  displayName?: string
-  //in case of a child component needed to use its parent's tokens, provide parent's name
-  useTokensFrom?: keyof NewComponentTypes
-}): ReturnType<GenerateStyle> => {
-  const { generateStyle, params, componentId, themeOverride } = useStyleParams
-  const useTokensFrom = useStyleParams.useTokensFrom
-  const themeInContext = useTheme() as Theme
-
-  const themeOverrideFromProvider = themeInContext.themeOverride
-  const componentWithTokensId =
-    useTokensFrom ?? (componentId as keyof NewComponentTypes)
-
-  // resolving the theming functions and applying the overrides
-  const primitiveOverrides = themeOverrideFromProvider?.primitives
-  const semanticsOverrides = themeOverrideFromProvider?.semantics
-  const sharedTokensOverrides = themeOverrideFromProvider?.sharedTokens
-  const componentOverridesFromSettingsProvider =
-    themeOverrideFromProvider?.components?.[componentWithTokensId]
-
-  const primitives = mergeDeep(
-    themeInContext.newTheme.primitives,
-    primitiveOverrides!
-  )
-
-  const semantics = mergeDeep(
-    themeInContext.newTheme.semantics?.(primitives),
-    semanticsOverrides!
-  )
-
-  const sharedTokens = mergeDeep(
-    themeInContext.newTheme.sharedTokens?.(semantics),
-    sharedTokensOverrides as Record<string, unknown>
-  )
+  P extends (componentTheme: any, params: any, theme: any) => any
+>(
+  useStyleParams: UseStyleParamsWithTheme<P> | UseStyleParamsWithoutTheme<P>
+): ReturnType<P> => {
+  const { generateStyle, params, componentId, displayName } = useStyleParams
+  const generateComponentTheme: GenerateComponentTheme = (
+    useStyleParams as UseStyleParamsWithTheme<P>
+  )?.generateComponentTheme
+  const theme = useTheme()
 
   const baseComponentTheme =
-    themeInContext.newTheme.components[componentWithTokensId]?.(semantics)
+    typeof generateComponentTheme === 'function'
+      ? generateComponentTheme(theme as Theme)
+      : {}
 
-  const componentThemeFromSettingsProvider = mergeDeep(
-    baseComponentTheme,
-    componentOverridesFromSettingsProvider as Record<string, unknown>
-  ) as ComponentTheme
-
-  const componentTheme = mergeDeep(
-    componentThemeFromSettingsProvider,
-    typeof themeOverride === 'function'
-      ? themeOverride(componentThemeFromSettingsProvider, themeInContext)
-      : themeOverride!
+  const themeOverride = getComponentThemeOverride(
+    theme,
+    displayName ?? componentId ?? '',
+    componentId,
+    params?.themeOverride,
+    baseComponentTheme
   )
 
-  // @ts-ignore TODO-theme-types: fix typing
-  return generateStyle(componentTheme, params, sharedTokens)
+  const componentTheme = { ...baseComponentTheme, ...themeOverride }
+
+  return generateStyle(componentTheme, params ? params : {}, theme)
 }
 
 export default useStyle
 export { useStyle }
-export type { NewThemeOverrideProp }
