+++ packages/ui-form-field/src/FormFieldLayout/v2/index.tsx
@@ -26,7 +26,7 @@ import { forwardRef, useEffect, useState, useCallback } from 'react'
 import { hasVisibleChildren } from '@instructure/ui-a11y-utils'
 import { omitProps, useDeterministicId } from '@instructure/ui-react-utils'
 
-import { useStyle } from '@instructure/emotion'
+import { useStyleNew } from '@instructure/emotion'
 import { FormFieldMessages } from '../../FormFieldMessages/v2'
 import generateStyle from './styles'
 import { allowedProps } from './props'
@@ -108,7 +108,7 @@ const FormFieldLayout = forwardRef<Element, FormFieldLayoutProps>(
     )
 
     // Styles
-    const styles = useStyle({
+    const styles = useStyleNew({
       generateStyle,
       themeOverride,
       params: {
