+++ packages/ui-instructure/src/DataPermissionLevels/v2/index.tsx
@@ -27,7 +27,7 @@ import { Button, CloseButton } from '@instructure/ui-buttons/latest'
 import { Heading } from '@instructure/ui-heading/latest'
 import { Text } from '@instructure/ui-text/latest'
 import { Link } from '@instructure/ui-link/latest'
-import { useStyle } from '@instructure/emotion'
+import { useStyleNew } from '@instructure/emotion'
 
 import { DataPermissionLevelsProps } from './props'
 import generateStyle from './styles'
@@ -51,7 +51,7 @@ const DataPermissionLevels = ({
 }: DataPermissionLevelsProps) => {
   const [open, setOpen] = useState(false)
 
-  const styles = useStyle({
+  const styles = useStyleNew({
     generateStyle,
     themeOverride,
     componentId: 'DataPermissionLevels',
