+++ packages/ui-instructure/src/NutritionFacts/v1/index.tsx
@@ -27,7 +27,7 @@ import { Button, CloseButton } from '@instructure/ui-buttons/v11_6'
 import { Heading } from '@instructure/ui-heading/v11_6'
 import { Text } from '@instructure/ui-text/v11_6'
 import { Link } from '@instructure/ui-link/v11_6'
-import { useStyleLegacy as useStyle } from '@instructure/emotion'
+import { useStyle as useStyleNew } from '@instructure/emotion'
 
 import { NutritionFactsProps } from './props'
 import generateStyle from './styles'
@@ -50,7 +50,7 @@ const NutritionFacts = ({
 }: NutritionFactsProps) => {
   const [open, setOpen] = useState(false)
 
-  const styles = useStyle({
+  const styles = useStyleNew({
     generateStyle,
     generateComponentTheme,
     componentId: 'NutritionFacts',
